refactor: align catalog review feedback in types and config migration
